+++ src/config/settings.ts
@@ -125,6 +125,13 @@ const settings: Settings = {
         ttl: parseInt(process.env.METRICS_REPORTING_LOCK_TTL || '60'),
       },
     },
+    blockchainMetrics: {
+      interval: parseInt(process.env.VALIDATORS_RESOLVER_JOB_INTERVAL || '600000'),
+      lock: {
+        name: process.env.VALIDATORS_RESOLVER_LOCK_NAME || 'lock::ValidatorsResolver',
+        ttl: parseInt(process.env.VALIDATORS_RESOLVER_LOCK_TTL || '60'),
+      },
+    },
     blockDispatcher: resolveBlockDispatcherSettings(),
   },
   redis: {
@@ -159,6 +166,9 @@ const settings: Settings = {
       chainId: ChainsIds.BSC,
     },
     contracts: {
+      bank: {
+        name: process.env.STAKING_BANK_NAME || 'StakingBank',
+      },
       chain: {
         name: 'Chain',
       },
@@ -243,6 +253,7 @@ const settings: Settings = {
   environment: process.env.ENVIRONMENT || process.env.NODE_ENV,
   name: process.env.NEW_RELIC_APP_NAME || process.env.NAME || 'default',
   deviationTrigger: {
+    heartbeatRounds: getTimeSetting(parseInt(process.env.DEVIATION_HEARTBEAT_ROUNDS || '5'), 3),
     roundLengthSeconds: getTimeSetting(parseInt(process.env.DEVIATION_ROUND_LENGTH || '60'), 30),
     leaderInterval: getTimeSetting(parseInt(process.env.DEVIATION_JOB_INTERVAL || '10000'), 5000),
     lock: {