+++ src/types/NumberOfSignatures.ts
@@ -1,0 +1,4 @@
+import {BlockchainType} from './Settings';
+
+export type NumberOfSignatures = Record<BlockchainType, number>;
+export type NumberOfSignaturesPerChain = Record<string, NumberOfSignatures>;